Relocate classes to package
Board, BoardTest, Pawn, PawnTest 클래스의 패키지를 재배치했습니다.
