+++ src/test/java/pieces/PawnTest.java
@@ -1,4 +1,4 @@
-package net.honux.chess;
+package pieces;
 
 import org.junit.jupiter.api.*;
 import static org.assertj.core.api.Assertions.*;
